chore: remove unnecessary shit before merge
