+++ contracts/actions/aave/v2/Withdraw.sol
@@ -7,7 +7,7 @@ import { OperationStorage } from "../../../core/OperationStorage.sol";
 import { ILendingPool } from "../../../interfaces/aave/ILendingPool.sol";
 import { WithdrawData } from "../../../core/types/Aave.sol";
 import { AAVE_LENDING_POOL, WITHDRAW_ACTION } from "../../../core/constants/Aave.sol";
-import "hardhat/console.sol";
+
 /**
  * @title Withdraw | AAVE Action contract
  * @notice Withdraw collateral from AAVE's lending pool
