+++ packages/oasis-actions/src/helpers/constants.ts
@@ -5,8 +5,6 @@ export const DEFAULT_FEE = 20
 export const FEE_BASE = 10000
 
 export const TYPICAL_PRECISION = 18
-export const DEFAULT_REBASE_OFFSET = 2
-
 export const NULL_ADDRESS = '0x0000000000000000000000000000000000000000'
 
 export const ZERO = new BigNumber(0)
