--- packages/oasis-actions/src/helpers/swap/isRebasingToken.ts
@@ -1,3 +1,0 @@
-export function isRebasingToken(tokenSymbol: string) {
-  return ['STETH'].includes(tokenSymbol)
-}
