+++ test/aave/Open.test.ts
@@ -104,89 +104,6 @@ describe(`Strategy | AAVE | Open Position`, async function () {
         })
       })
     })
-    describe('Open position: With 1inch', function () {
-      before(async () => {
-        fixture = await loadFixture(getSystemWithAavePositions({ use1inch: true }))
-      })
-
-      describe('Using DSProxy', () => {
-        let position: IPosition
-        let simulatedPosition: IPosition
-        let simulatedTransition: IPositionTransition['simulation']
-        let feeWalletBalanceChange: BigNumber
-
-        before(async () => {
-          const { dsProxyPosition: dsProxyStEthEthEarnPositionDetails } = fixture
-          position = await dsProxyStEthEthEarnPositionDetails.getPosition()
-          simulatedPosition = dsProxyStEthEthEarnPositionDetails.__openPositionSimulation.position
-          simulatedTransition = dsProxyStEthEthEarnPositionDetails.__openPositionSimulation
-          feeWalletBalanceChange = dsProxyStEthEthEarnPositionDetails.__feeWalletBalanceChange
-        })
-
-        it('Should draw the correct amount of debt', async () => {
-          expectToBe(
-            simulatedPosition.debt.amount.toFixed(0),
-            'lte',
-            position.debt.amount.toFixed(0),
-          )
-        })
-        it('Should deposit all collateral', async () => {
-          expectToBe(
-            simulatedPosition.collateral.amount,
-            'lte',
-            position.collateral.amount.toFixed(0),
-          )
-        })
-        it('Should have the correct multiple', async () => {
-          expectToBe(position.riskRatio.multiple, 'lte', simulatedPosition.riskRatio.multiple)
-        })
-        it('Should collect fee', async () => {
-          expectToBeEqual(simulatedTransition.swap.tokenFee, feeWalletBalanceChange)
-        })
-      })
-      describe('Using DPM Proxy', async () => {
-        supportedStrategies.forEach(({ name: strategy }) => {
-          let position: IPosition
-          let simulatedPosition: IPosition
-          let simulatedTransition: IPositionTransition['simulation']
-          let feeWalletBalanceChange: BigNumber
-
-          before(async function () {
-            const { dpmPositions } = fixture
-            const positionDetails = dpmPositions[strategy]
-
-            if (!positionDetails) {
-              this.skip()
-            }
-            position = await positionDetails.getPosition()
-            simulatedPosition = positionDetails.__openPositionSimulation.position
-            simulatedTransition = positionDetails.__openPositionSimulation
-            feeWalletBalanceChange = positionDetails.__feeWalletBalanceChange
-          })
-
-          it(`Should draw the correct amount of debt for ${strategy}`, async () => {
-            expectToBe(
-              simulatedPosition.debt.amount.toFixed(0),
-              'lte',
-              position.debt.amount.toFixed(0),
-            )
-          })
-          it(`Should deposit all collateral for ${strategy}`, async () => {
-            expectToBe(
-              simulatedPosition.collateral.amount,
-              'lte',
-              position.collateral.amount.toFixed(0),
-            )
-          })
-          it(`Should have the correct multiple for ${strategy}`, async () => {
-            expectToBe(position.riskRatio.multiple, 'lte', simulatedPosition.riskRatio.multiple)
-          })
-          it(`Should collect fee for ${strategy}`, async () => {
-            expectToBeEqual(simulatedTransition.swap.tokenFee, feeWalletBalanceChange)
-          })
-        })
-      })
-    })
   })
   describe('Using AAVE V3', async function () {
     let fixture: SystemWithAAVEV3Positions
